requirements.txt
# Основные зависимости
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.1

# Тестирование
pytest==8.0.2
pytest-asyncio==0.23.5

# Утилиты
pydantic==2.6.1 